+++ src/pages/Articles/components/FormCategory/FormCategory.module.scss
@@ -5,6 +5,10 @@
   justify-content: space-between;
   flex-wrap: wrap;
 
+  @include media.media-breakpoint(mobile) {
+    flex-direction: column;
+  }
+
   &__wrapper {
     position: relative;
     display: flex;
@@ -85,6 +89,8 @@
 
   @include media.media-breakpoint(mobile) {
     left: initial;
+    bottom: initial;
     right: 0;
+    top: -53px;
   }
 }
